Refactor accounts
